requirements.txt
django
djangorestframework
psycopg2
django-cors-headers
djangorestframework-simplejwt
python-slugify
pillow
requests